generator runs on the entire Sail-RISCV ast definition and generates a C struct successfully

+++ lib/decode_procedure.ml
@@ -1,7 +1,7 @@
 open Decoder
 
 type bv_val = string
-type bv_index = int64
+type bv_index = int
 type enum_val = string
 
 type bv_expr =
